Revert "Removed commit a61d4143c7f1a8607fec62bb07c3fc1c264e6e4c"

+++ lib/dailyploy/schema/task.ex
@@ -19,7 +19,7 @@ defmodule Dailyploy.Schema.Task do
     field :start_datetime, :utc_datetime
     field :end_datetime, :utc_datetime
     field :comments, :string
-    field :estimation, :float
+    field :estimation, :integer
     field :status, :string, default: "not_started"
     field :priority, :string
     field :is_complete, :boolean, default: false, null: false
